+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for all transactions relevant from protocol version 7 and onwards. This includes expanding `AccountTransactionPayload`
   with the necessary variants, and corresponding constructor functions for `AccountTransaction`.
 - `WalletConnectSendTransactionParam` and `WalletConnectSignMessageParam` for decoding parameters received with walletconnect requests.
+- `ContractSchema`, `TypeSchema`, and `ModuleSchema` for encoding/decoding data from/to the corresponding JSON representation
 
 #### GRPC client
 - `NodeClient.status` to query transaction status.